+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 -   [ENH] `clean_names` can now be applied to column values. Issue #995 @samukweku
+-   [BUG] Fix ImportError - Issue #1285 @samukweku
 
 ## [v0.25.0] - 2023-07-27